Integracion con yama (archivoInfo). Ajustes en validaciones de comandos cpfrom y mkdir. Nuevos nodos se guardan en nodos.bin (parte de la inicializacion en proceso)

+++ filesystem/src/estructuras.h
@@ -14,7 +14,7 @@
 typedef struct
 	__attribute__((packed)) {
 		char idMensaje;
-		long size;
+		int size;
 	} t_header;
 
 	typedef struct {
